Deprecate setTraceEntity
